Rename some files and add identicon testcases
This commit will not pass tests as identicons tests are implemented but not
the actuall logic to create them.
